feat: Move static buildQuery method into Driver.
Also makes the Driver interface generic on the query and transaction
types.
